+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/TemporaryAnnotations.kt
@@ -11,3 +11,7 @@ annotation class UselessCallOnNotNullType
 /**暗示智能类型转换，但是无法适用的扩展方法。*/
 @Retention(SOURCE)
 annotation class ImpliesSmartCast
+
+/**有待以自动生成的方式编写的代码。*/
+@Retention(SOURCE)
+annotation class ToBeGeneratedCode