+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/domain/DataEntity.kt
@@ -1,0 +1,9 @@
+package com.windea.breezeframework.core.domain
+
+/**
+ * 数据实体。
+ *
+ * 数据实体用于存储一系列的数据，并且保证最小的可见性。
+ * 数据实体应当是一个数据类，可以通过其构造方法创建，可以通过其构建器的构建方法或代码块创建。
+ * */
+interface DataEntity