+++ breeze-mapper/src/main/kotlin/com/windea/breezeframework/mapper/impl/PropertiesMapper.kt
@@ -1,0 +1,25 @@
+package com.windea.breezeframework.mapper.impl
+
+import com.windea.breezeframework.mapper.*
+
+class PropertiesMapper : Mapper {
+	override fun <T> unmap(string: String, type: Class<T>): T {
+		TODO("not implemented")
+	}
+
+	override fun <T> map(data: T): String {
+		TODO("not implemented")
+	}
+
+	private fun Map<*, *>.map() {
+		TODO()
+	}
+
+	private fun List<*>.map() {
+		TODO()
+	}
+
+	private fun Any?.defaultMap() {
+		TODO()
+	}
+}