+++ breeze-mapper/src/main/kotlin/com/windea/breezeframework/mapper/impl/XmlMapper.kt
@@ -1,0 +1,13 @@
+package com.windea.breezeframework.mapper.impl
+
+import com.windea.breezeframework.mapper.*
+
+class XmlMapper : Mapper {
+	override fun <T> unmap(string: String, type: Class<T>): T {
+		TODO("not implemented")
+	}
+
+	override fun <T> map(data: T): String {
+		TODO("not implemented")
+	}
+}