+++ breeze-unstable/src/main/kotlin/com/windea/breezeframework/core/domain/EnumConstant.kt
@@ -1,4 +1,8 @@
 package com.windea.breezeframework.core.domain
 
-/**声明有可枚举的常量的类或单例对象。*/
+/**
+ * 枚举常量的接口。
+ *
+ * 这个接口的实现应当是单例对象，枚举了一系列的同一类型的常量。
+ */
 interface EnumConstant<T>